backend/requirements/base.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-storages[s3]>=1.14,<2.0
boto3>=1.35,<2.0
django-redis>=5.4,<6.0
celery>=5.4,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
Pillow>=10.4,<11.0
psycopg[binary]>=3.2,<4.0
python-decouple>=3.8,<4.0
gunicorn>=22.0,<23.0
